requirements.txt
# ============================================================
# TimeWoven — Python Dependencies
# ============================================================
# Дата обновления: 2026-04-21
# Python: 3.11+
# Окружение: Production + Development
#
# Установка:
#   pip install -r requirements.txt
#
# Обновление после добавления пакета:
#   pip freeze > requirements.lock.txt   # полный снимок
#   Вручную синхронизировать этот файл
#
# Конвенция версионирования:
#   package==X.Y.Z   — точная версия (для стабильности прода)
#   package>=X.Y,<Z  — диапазон (для гибкости при разработке)
# ============================================================


# --- Веб-фреймворк и ASGI ---
fastapi==0.128.8             # REST API фреймворк, маршрутизация, валидация
uvicorn==0.39.0              # ASGI-сервер для запуска FastAPI


# --- ORM и база данных ---
SQLAlchemy==2.0.49           # ORM и query layer
psycopg2-binary==2.9.11      # PostgreSQL адаптер для SQLAlchemy


# --- Шаблонизация и UI ---
Jinja2==3.1.6                # Серверный рендеринг HTML-шаблонов
MarkupSafe==3.0.3            # Экранирование в шаблонах Jinja2
python-multipart==0.0.20     # Обработка form-data (загрузка файлов)


# --- Конфигурация и сессии ---
python-dotenv==1.2.2         # Загрузка переменных окружения из .env
itsdangerous==2.2.0          # Подпись и верификация cookies/токенов
starlette-session==0.4.3     # Сессии для Starlette/FastAPI


# --- Валидация и сериализация ---
pydantic==2.12.5             # Валидация данных, схемы запросов/ответов
pydantic_core==2.41.5        # Ядро pydantic v2
annotated-types==0.7.0       # Дополнительные типовые аннотации для pydantic


# --- Async и низкоуровневые зависимости ---
anyio==4.12.1                # Абстракция над asyncio/Trio
httpx==0.28.1               # HTTP-клиент для webhook/bot/local_stub интеграций
h11==0.16.0                  # HTTP/1.1 протокол для uvicorn
idna==3.11                   # Поддержка IDN-доменов
click==8.1.8                 # CLI-хэлпер (используется зависимостями)
exceptiongroup==1.3.1        # Группировка исключений (для совместимости)


# --- Утилиты типов ---
typing_extensions==4.15.0    # Дополнительные typing-фичи
typing-inspection==0.4.2     # Инспекция типов (вспомогательные утилиты)
annotated-doc==0.0.4         # Работа с аннотированной документацией


# --- Тестирование (dev only, TODO) ---
# pytest==8.1.1              # Тестовый фреймворк
# pytest-cov==5.0.0          # Покрытие кода тестами


# --- Линтинг и форматирование (dev only, TODO) ---
# ruff==0.3.5                # Линтер и форматтер


# --- AI / ML и интеграции (опционально, TODO) ---
anthropic==0.96.0            # Claude API (Anthropic) для NLP-анализа воспоминаний
# openai-whisper==20231117   # Транскрибация аудио (Whisper pipeline)
# torch>=2.0,<3.0            # PyTorch — зависимость Whisper
# python-telegram-bot==21.1  # Telegram Bot (Импульс дня)